feat(capacity): upgrade VRAM tracking from u32 to u64
AtomicU32/u32 capped at ~4.3 PB; AtomicU64/u64 supports up to ~18.4 EB.
Updates VramPoolPort trait, VramPool, DistributedVramPool, analyzer,
placement_planner, and dashboard_handlers to use u64 throughout.

+++ crates/veronex/src/infrastructure/inbound/http/dashboard_handlers.rs
@@ -182,9 +182,9 @@ pub struct LoadedModelInfo {
 pub struct ProviderVramInfo {
     pub provider_id:     String,
     pub provider_name:   String,
-    pub total_vram_mb:   u32,
-    pub used_vram_mb:    u32,
-    pub available_vram_mb: u32,
+    pub total_vram_mb:   u64,
+    pub used_vram_mb:    u64,
+    pub available_vram_mb: u64,
     pub thermal_state:   String,
     pub temp_c:          Option<f32>,
     pub loaded_models:   Vec<LoadedModelInfo>,
